new folder structure app folder would contain stacked setup
