+++ lib/main.dart
@@ -1,7 +1,6 @@
-import 'package:ellen/ui/view/dashboard/dashboard.dart';
-import 'package:ellen/ui/view/login/login.dart';
-import 'package:ellen/ui/view/onboarding/onboarding.dart';
 import 'package:flutter/material.dart';
+
+import 'ui/screens/splash_screen/splash_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -19,7 +18,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const OnBoarding(),
+      home: const SplashScreen(),
     );
   }
 }
