+++ lib/ui/screens/dashboard/dashboard.dart
@@ -1,11 +1,12 @@
-import 'package:ellen/ui/view/dashboard/sub_view/account.dart';
-import 'package:ellen/ui/view/dashboard/sub_view/buy.dart';
-import 'package:ellen/ui/view/market/market.dart';
-import 'package:ellen/ui/view/dashboard/sub_view/save.dart';
-import 'package:ellen/ui/view/dashboard/sub_view/wallet.dart';
+import 'package:ellen/ui/screens/dashboard/sub_view/account.dart';
+import 'package:ellen/ui/screens/dashboard/sub_view/buy.dart';
+import 'package:ellen/ui/screens/dashboard/sub_view/save.dart';
+import 'package:ellen/ui/screens/dashboard/sub_view/wallet.dart';
 import 'package:ellen/utils/colors.dart';
 import 'package:ellen/utils/styles.dart';
 import 'package:flutter/material.dart';
+
+import '../market/market.dart';
 
 class DashBoard extends StatefulWidget {
   const DashBoard({Key? key}) : super(key: key);
