+++ lib/ui/screens/market/view_model/market_view_model.dart
