+++ lib/ui/screens/onboarding/widgets/onboarding_overlay.dart
